mini games
I just created a small sequencer game for fun. You just remember the sequence and click the boxes in the correct order. I added music which was my way of learning some things about audio in html. It's pretty fun to play when I've been coding for a long time and need a mental break. I know it sounds crazy because I'm using my mind to remember the sequence, but I guess the music is so groovy that I don't mind.