senior capstone
For my senior capstone, me and my group created a new and improved patient portal. One of our deliverables was image authentication. The security page in this folder shows how that authentication looks. A user would simply choose the picture they wanted, and them select that same picture when prompted to login as an existing user. The image selection pool consists of the images with the "imageauth" file path. The sign up page was designed as the page the users would sign up on. I've also included a "portalux.png" file as a snippet of my designing. I created that logo using Microsoft Word. I've also included index2 and index3, which are javascript files. These hold the code that allows there to be communication between the forms and the databases. I used MySQL Workbench to create the database tables, however that link is not included here.